+++ library/core/src/main/java/com/google/android/exoplayer2/text/SimpleSubtitleDecoder.java
@@ -55,17 +55,12 @@ public abstract class SimpleSubtitleDecoder extends
 
   @Override
   protected final SubtitleOutputBuffer createOutputBuffer() {
-    return new SimpleSubtitleOutputBuffer(this);
+    return new SimpleSubtitleOutputBuffer(this::releaseOutputBuffer);
   }
 
   @Override
   protected final SubtitleDecoderException createUnexpectedDecodeException(Throwable error) {
     return new SubtitleDecoderException("Unexpected decode error", error);
-  }
-
-  @Override
-  protected final void releaseOutputBuffer(SubtitleOutputBuffer buffer) {
-    super.releaseOutputBuffer(buffer);
   }
 
   @SuppressWarnings("ByteBufferBackingArray")